# go-architecture/hexagonal-architecture/go-hexagonal-architecture-2/docs/swagger.yaml
basePath: /v1
definitions:
...
  handler.categoryResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Foods
        type: string
    type: object
...
  handler.createProductRequest:
    properties:
      category_id:
        example: 1
        minimum: 1
        type: integer
      image:
        example: https://example.com/chiki-ball.png
        type: string
      name:
        example: Chiki Ball
        type: string
      price:
        example: 5000
        minimum: 0
        type: number
      stock:
        example: 100
        minimum: 0
        type: integer
    required:
    - category_id
    - image
    - name
    - price
    - stock
    type: object
  handler.errorResponse:
    properties:
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handler.meta:
    properties:
      limit:
        example: 10
        type: integer
      skip:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  handler.productResponse:
    properties:
      category:
        $ref: '#/definitions/handler.categoryResponse'
      created_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/chiki-ball.png
        type: string
      name:
        example: Chiki Ball
        type: string
      price:
        example: 5000
        type: number
      sku:
        example: 9a4c25d3-9786-492c-b084-85cb75c1ee3e
        type: string
      stock:
        example: 100
        type: integer
      updated_at:
        example: "1970-01-01T00:00:00Z"
        type: string
    type: object
...
  handler.response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  handler.updateProductRequest:
    properties:
      category_id:
        example: 1
        minimum: 1
        type: integer
      image:
        example: https://example.com/nutrisari-jeruk.png
        type: string
      name:
        example: Nutrisari Jeruk
        type: string
      price:
        example: 2000
        minimum: 0
        type: number
      stock:
        example: 200
        minimum: 0
        type: integer
    required:
    - category_id
    - image
    - name
    - price
    - stock
    type: object
...
host: gopos.bagashiz.me
info:
  contact:
    name: test
    url: test
  description: This is a simple RESTful Point of Sale (POS) Service API written in
    Go using Gin web framework, PostgreSQL database, and Redis cache.
  license:
    name: MIT
    url: test
  title: Go POS (Point of Sale) API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: List products with pagination
      parameters:
      - description: Category ID
        in: query
        name: category_id
        type: integer
      - description: Query
        in: query
        name: q
        type: string
      - description: Skip
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved
          schema:
            $ref: '#/definitions/handler.meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: create a new product with name, image, price, and stock
      parameters:
      - description: Create product request
        in: body
        name: createProductRequest
        required: true
        schema:
          $ref: '#/definitions/handler.createProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product created
          schema:
            $ref: '#/definitions/handler.productResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted
          schema:
            $ref: '#/definitions/handler.response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: get a product by id with its category
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved
          schema:
            $ref: '#/definitions/handler.productResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: update a product's name, image, price, or stock by id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product request
        in: body
        name: updateProductRequest
        required: true
        schema:
          $ref: '#/definitions/handler.updateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated
          schema:
            $ref: '#/definitions/handler.productResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"